Fix format warning in Tuya component
Fix the following compile time warning in the Tuya component:
src/esphome/core/log.h:110:99: warning: format '%hhX' expects argument of type 'int', but argument 6 has type 'esphome::tuya::TuyaDatapointType' [-Wformat=]
   esp_log_printf_(ESPHOME_LOG_LEVEL_WARN, tag, __LINE__, ESPHOME_LOG_FORMAT(format), ##__VA_ARGS__)
                                                                                                   ^

+++ esphome/components/tuya/tuya.cpp
@@ -288,7 +288,7 @@ void Tuya::handle_datapoint_(const uint8_t *buffer, size_t len) {
       ESP_LOGD(TAG, "Datapoint %u update to %#08X", datapoint.id, datapoint.value_bitmask);
       break;
     default:
-      ESP_LOGW(TAG, "Datapoint %u has unknown type %#02hhX", datapoint.id, datapoint.type);
+      ESP_LOGW(TAG, "Datapoint %u has unknown type %#02hhX", datapoint.id, static_cast<int>(datapoint.type));
       return;
   }
